requirements.txt
discord.py
aiosqlite
python-dotenv
pytz
numpy
pillow
anyio